requirements.txt
Django==3.2
djangorestframework==3.13.1
django-dotenv==1.4.2
ipdb==0.13.9
ipython==7.31.0
jedi==0.18.1
psycopg2==2.9.3
